Broker_Streamer_Server/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.4.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.1.1
click==8.1.8
crewai==1.14.3
cryptography==47.0.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.18.0
durationpy==0.10
et_xmlfile==2.0.0
fastuuid==0.14.0
filelock==3.29.0
Flask==3.1.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.0
idna==3.13
importlib_metadata==8.5.0
importlib_resources==7.1.0
influxdb-client==1.50.0
instructor==1.15.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
json5==0.10.0
json_repair==0.25.3
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lance-namespace==0.7.2
lance-namespace-urllib3-client==0.7.2
lancedb==0.30.0
linkify-it-py==2.1.0
litellm==1.83.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.25.1
openai==2.24.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.8
overrides==7.7.0
packaging==26.2
paho-mqtt==2.1.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
platformdirs==4.9.6
portalocker==2.7.0
posthog==5.4.0
propcache==0.4.1
protobuf==5.29.6
pyarrow==24.0.0
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.20.0
PyJWT==2.12.1
pypdfium2==5.7.1
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
reactivex==4.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.4
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.4.1
starlette==1.0.0
tenacity==9.1.4
textual==8.2.4
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.67.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
urllib3==2.6.3
uv==0.11.8
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.8
wheel==0.47.0
yarl==1.23.0
zipp==3.23.1
